+++ Controllers/TransactionController.cs
@@ -1,21 +1,35 @@
 using Captain.DTOs;
 using Captain.Services;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Captain.controllers;
 
 [Route("api/[controller]")]
 [ApiController]
-public class TransactionController(ITransactionService transactionService) : ControllerBase
+[Authorize]
+public class TransactionController(
+  ITransactionService transactionService,
+  UserManager<AppUser> userManager
+) : ControllerBase
 {
   private readonly ITransactionService _transactionService = transactionService;
+  private readonly UserManager<AppUser> _userManager = userManager;
+  private string UserId => _userManager.GetUserId(User)!;
 
   [HttpGet("get-all")]
   public async Task<ActionResult<TransactionResponse>> GetTransactionsAsync(
     CancellationToken cancellationToken
   )
   {
+    if (UserId is null)
+    {
+      return Unauthorized();
+    }
+
     var transactions = await _transactionService.GetTransactionsAsync(
+      userId: UserId,
       cancellationToken: cancellationToken
     );
     return Ok(transactions);
@@ -27,7 +41,13 @@ public class TransactionController(ITransactionService transactionService) : Con
     CancellationToken cancellationToken
   )
   {
+    if (UserId is null)
+    {
+      return Unauthorized();
+    }
+
     var transaction = await _transactionService.GetTransactionByIdAsync(
+      userId: UserId,
       transactionId,
       cancellationToken
     );
@@ -46,7 +66,13 @@ public class TransactionController(ITransactionService transactionService) : Con
     CancellationToken cancellationToken
   )
   {
+    if (UserId is null)
+    {
+      return Unauthorized();
+    }
+
     var createdTransaction = await _transactionService.CreateTransactionAsync(
+      userId: UserId,
       request: request,
       cancellationToken: cancellationToken
     );
@@ -60,7 +86,13 @@ public class TransactionController(ITransactionService transactionService) : Con
     CancellationToken cancellationToken
   )
   {
+    if (UserId is null)
+    {
+      return Unauthorized();
+    }
+
     var updatedTransaction = await _transactionService.UpdateTransactionAsync(
+      userId: UserId,
       request: request,
       cancellationToken: cancellationToken
     );
@@ -78,7 +110,13 @@ public class TransactionController(ITransactionService transactionService) : Con
     CancellationToken cancellationToken
   )
   {
+    if (UserId is null)
+    {
+      return Unauthorized();
+    }
+
     var response = await _transactionService.DeleteTransactionAsync(
+      userId: UserId,
       transactionId: transactionId,
       cancellationToken: cancellationToken
     );
